+++ src/com/kaylerrenslow/armaDialogCreator/arma/header/HeaderParser.java
@@ -1,13 +1,15 @@
 package com.kaylerrenslow.armaDialogCreator.arma.header;
 
 import com.kaylerrenslow.armaDialogCreator.util.CharSequenceReader;
-import org.antlr.v4.runtime.ANTLRInputStream;
-import org.antlr.v4.runtime.CommonTokenStream;
+import org.antlr.v4.runtime.*;
+import org.antlr.v4.runtime.atn.ATNConfigSet;
+import org.antlr.v4.runtime.dfa.DFA;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.util.BitSet;
 import java.util.List;
 import java.util.ResourceBundle;
 
@@ -63,15 +65,10 @@ public class HeaderParser {
 	private HeaderFile doParse() throws Exception {
 		HeaderFile headerFile = new HeaderFile(getParsingFile());
 
-		PreprocessCallback callback = new PreprocessCallback() {
-			@Override
-			public void fileProcessed(@NotNull File file, @NotNull Preprocessor.PreprocessorInputStream fileContentStream) throws HeaderParseException {
-				parseText(headerFile, fileContentStream);
-			}
-		};
+		Preprocessor pre = new Preprocessor(parsingFile, parserContext);
+		Preprocessor.PreprocessorInputStream fileContentStream = pre.run();
 
-		Preprocessor pre = new Preprocessor(parsingFile, parserContext, callback);
-		pre.run();
+		parseText(headerFile, fileContentStream);
 
 		return headerFile;
 	}
@@ -82,6 +79,7 @@ public class HeaderParser {
 		l.getErrorListeners().clear();
 		p.getErrorListeners().clear();
 
+		p.addErrorListener(HeaderParserErrorListener.INSTANCE);
 
 		HeaderClass rootClass = p.root_class().ast;
 
@@ -115,9 +113,44 @@ public class HeaderParser {
 		try {
 			return new HeaderAntlrLexer(new ANTLRInputStream(s));
 		} catch (IOException e) {
-			throw new IllegalStateException(e);
+			throw new RuntimeException(e);
 		}
 	}
 
 
+	public static class HeaderParserErrorListener extends BaseErrorListener {
+		public static final HeaderParserErrorListener INSTANCE = new HeaderParserErrorListener();
+
+		@Override
+		public void syntaxError(Recognizer<?, ?> recognizer, Object offendingSymbol, int line, int charPositionInLine, String msg, RecognitionException e) {
+			//			String sourceName = recognizer.getInputStream().getSourceName();
+			//			System.out.println("HeaderParserErrorListener.syntaxError offendingSymbol=" + offendingSymbol);
+			//			System.out.println("HeaderParserErrorListener.syntaxError offendingSymbol.getClass()=" + offendingSymbol.getClass());
+			//			CommonToken t = (CommonToken) offendingSymbol;
+			//
+			//			if (!sourceName.isEmpty()) {
+			//				sourceName = String.format("%s:%d:%d: ", sourceName, line, charPositionInLine);
+			//			}
+			//
+			//			System.err.println(sourceName+"line "+line+":"+charPositionInLine+" "+msg);
+		}
+
+		@Override
+		public void reportAmbiguity(Parser recognizer, DFA dfa, int startIndex, int stopIndex, boolean exact, BitSet ambigAlts, ATNConfigSet configs) {
+			super.reportAmbiguity(recognizer, dfa, startIndex, stopIndex, exact, ambigAlts, configs);
+			System.out.println("HeaderParserErrorListener.reportAmbiguity");
+		}
+
+		@Override
+		public void reportAttemptingFullContext(Parser recognizer, DFA dfa, int startIndex, int stopIndex, BitSet conflictingAlts, ATNConfigSet configs) {
+			super.reportAttemptingFullContext(recognizer, dfa, startIndex, stopIndex, conflictingAlts, configs);
+			System.out.println("HeaderParserErrorListener.reportAttemptingFullContext");
+		}
+
+		@Override
+		public void reportContextSensitivity(Parser recognizer, DFA dfa, int startIndex, int stopIndex, int prediction, ATNConfigSet configs) {
+			super.reportContextSensitivity(recognizer, dfa, startIndex, stopIndex, prediction, configs);
+			System.out.println("HeaderParserErrorListener.reportContextSensitivity");
+		}
+	}
 }